requirements.txt
torch==2.10.0
torchvision==0.25.0
streamlit
pillow
matplotlib
numpy